refactor: move tests in tests directory
