+++ .github/workflows/pull-request.yml
@@ -76,4 +76,4 @@ jobs:
       - name: Run tests
         run: |
           source .venv/bin/activate
-          pytest tdp
+          pytest tests
