+++ pyproject.toml
@@ -70,3 +70,6 @@ select = ["F", "W", "I"]
 
 [tool.ruff.per-file-ignores]
 "**/__init__.py" = ["F401"]
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
