+++ test_dag_order/conftest.py
@@ -18,11 +18,11 @@ from tdp.core.collections import Collections
 from tdp.core.constants import YML_EXTENSION
 from tdp.core.dag import Dag
 from tdp.core.deployment import DeploymentRunner
-from tdp.core.deployment.test_deployment_runner import MockExecutor
 from tdp.core.models import DeploymentModel, init_database
 from tdp.core.operation import Operation
 from tdp.core.variables import ClusterVariables
 from tdp.dao import Dao
+from tests.unit.core.deployment.test_deployment_runner import MockExecutor
 
 from .constants import RULES_KEYS
 from .helpers import append_collection_list_action
