+++ tests/e2e/__init__.py
@@ -1,0 +1,2 @@
+# Copyright 2022 TOSIT.IO
+# SPDX-License-Identifier: Apache-2.0
