+++ tests/e2e/conftest.py
@@ -10,8 +10,8 @@ from click.testing import CliRunner
 from sqlalchemy import create_engine
 
 from tdp.cli.commands.init import init
-from tdp.conftest import generate_collection_at_path
 from tdp.core.models import BaseModel
+from tests.conftest import generate_collection_at_path
 
 
 class TDPInitArgs(NamedTuple):
