+++ tests/e2e/test_tdp_plan_dag.py
@@ -4,26 +4,21 @@
 
 from click.testing import CliRunner
 
-from tdp.cli.commands.conftest import TDPInitArgs
-from tdp.cli.commands.status.edit import edit
+from tdp.cli.commands.plan.dag import dag
+from tests.e2e.conftest import TDPInitArgs
 
 
-def test_tdp_status_edit(
+def test_tdp_plan_dag(
     tdp_init: TDPInitArgs,
 ):
     runner = CliRunner()
     result = runner.invoke(
-        edit,
+        dag,
         [
             "--collection-path",
             str(tdp_init.collection_path),
             "--database-dsn",
             tdp_init.db_dsn,
-            "--vars",
-            str(tdp_init.vars),
-            "service",
-            "--host",
-            "localhost",
         ],
     )
     assert result.exit_code == 0, result.output
