+++ tests/e2e/test_tdp_plan_ops.py
@@ -4,21 +4,22 @@
 
 from click.testing import CliRunner
 
-from tdp.cli.commands.conftest import TDPInitArgs
-from tdp.cli.commands.plan.dag import dag
+from tdp.cli.commands.plan.ops import ops
+from tests.e2e.conftest import TDPInitArgs
 
 
-def test_tdp_plan_dag(
+def test_tdp_plan_run(
     tdp_init: TDPInitArgs,
 ):
     runner = CliRunner()
     result = runner.invoke(
-        dag,
+        ops,
         [
             "--collection-path",
             str(tdp_init.collection_path),
             "--database-dsn",
             tdp_init.db_dsn,
+            "service_install",
         ],
     )
     assert result.exit_code == 0, result.output
