+++ tests/e2e/test_tdp_plan_reconfigure.py
@@ -4,22 +4,23 @@
 
 from click.testing import CliRunner
 
-from tdp.cli.commands.conftest import TDPInitArgs
-from tdp.cli.commands.plan.ops import ops
+from tdp.cli.commands.plan.reconfigure import reconfigure
+from tests.e2e.conftest import TDPInitArgs
 
 
-def test_tdp_plan_run(
+def test_tdp_plan_reconfigure(
     tdp_init: TDPInitArgs,
 ):
     runner = CliRunner()
     result = runner.invoke(
-        ops,
+        reconfigure,
         [
             "--collection-path",
             str(tdp_init.collection_path),
             "--database-dsn",
             tdp_init.db_dsn,
-            "service_install",
         ],
     )
-    assert result.exit_code == 0, result.output
+    assert (
+        result.exit_code == 1
+    ), result.output  # No stale components, hence nothing to reconfigure.
