+++ tests/e2e/test_tdp_status_edit.py
@@ -4,8 +4,8 @@
 
 from click.testing import CliRunner
 
-from tdp.cli.commands.conftest import TDPInitArgs
-from tdp.cli.commands.status.show import show
+from tdp.cli.commands.status.edit import edit
+from tests.e2e.conftest import TDPInitArgs
 
 
 def test_tdp_status_edit(
@@ -13,7 +13,7 @@ def test_tdp_status_edit(
 ):
     runner = CliRunner()
     result = runner.invoke(
-        show,
+        edit,
         [
             "--collection-path",
             str(tdp_init.collection_path),
@@ -21,6 +21,9 @@ def test_tdp_status_edit(
             tdp_init.db_dsn,
             "--vars",
             str(tdp_init.vars),
+            "service",
+            "--host",
+            "localhost",
         ],
     )
     assert result.exit_code == 0, result.output
