+++ tests/e2e/test_tdp_vars_validate.py
