+++ tests/unit/core/conftest.py
@@ -3,12 +3,12 @@
 
 import pytest
 
-from tdp.conftest import generate_collection_at_path
 from tdp.core.cluster_status import ClusterStatus
 from tdp.core.collection import Collection
 from tdp.core.collections import Collections
 from tdp.core.dag import Dag
 from tdp.core.variables import ClusterVariables
+from tests.conftest import generate_collection_at_path
 
 
 @pytest.fixture(scope="session")
