+++ tests/unit/core/entities/test_operation.py
