+++ tests/unit/core/test_baseenum.py
